// types/astrology.types.ts
export interface BirthData {
  nom: string;
  prenoms: string;
  genre: 'homme' | 'femme' | 'autre';
  dateNaissance: string; // Format: YYYY-MM-DD
  heureNaissance: string; // Format: HH:MM
  paysNaissance: string;
  villeNaissance: string;
  email?: string; // Email pour les notifications
}